requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-filter==25.2
Pillow==12.2.0
python-decouple==3.8
psycopg2-binary==2.9.12
